Relocate xelus/pachi/rgb/rev2 VIA logic
