+++ keyboards/xelus/pachi/rgb/rev2/keymaps/via/rules.mk
@@ -1,0 +1,1 @@
+VIA_ENABLE = yes
